sdk assignment 1 websearch/requirements.txt
openai>=1.0.0
tavily-python>=0.1.0
